Moved params over to params module and added euclid_param_get fn
