+++ src/params.h
@@ -8,6 +8,7 @@ extern "C" {
 #include "config.h"
 
 #include <stdint.h>
+#include <string.h>
 
 /* Parameters are variables (stored as individual bytes) which are
  * programmatically tracked, read from and written to EEPROM, and logged. Each
@@ -26,6 +27,30 @@ static const uint8_t mode_num_params[NUM_MODES] = {
     EUCLID_NUM_PARAMS, // EUCLID
 };
 
+#define PARAM_FLAGS_NONE 0x0
+#define PARAM_FLAG_MODIFIED 0x1
+#define PARAM_FLAG_NEEDS_WRITE 0x2
+
+/// Maximum size of `ParamsRuntime`'s tables. Must be large enough to store the
+/// `ParamId` type for any mode.
+#define PARAMS_RUNTIME_MAX 9
+
+/// Parameter properties which need to be modified at runtime. Each table has
+/// the same length (`.len`), and they are indexed by a mode's associated
+/// `ParamId` type.
+typedef struct ParamsRuntime {
+	/// Number of elements in tables
+	uint8_t len;
+	/// List of parameter values of length `.len`. The values are always assumed to be in bounds.
+	uint8_t values[PARAMS_RUNTIME_MAX];
+	/// List of parameter properties, stored as bitflags, of length `.len`. Bitflags are indexed
+	/// via `PARAM_FLAG_*` defines.
+	uint8_t flags[PARAMS_RUNTIME_MAX];
+} ParamsRuntime;
+
+/// Stores the runtime information of the active mode's parameters.
+ParamsRuntime params;
+
 /*
 Original EEPROM Schema:
 Channel 1: length = 1 density = 2 offset = 7
@@ -37,15 +62,7 @@ Channel 3: length = 5 density = 6 offset = 9
 /// backwards-compatibility with the original Sebsongs Euclidean firmware.
 static const Address euclid_param_addresses[EUCLID_NUM_PARAMS] = {1, 2, 7, 3, 4, 8, 5, 6, 9};
 
-Address param_address(Mode mode, ParamIdx idx) {
-	Address result = 0;
-	switch (mode) {
-		case MODE_EUCLID:
-			result = euclid_param_addresses[idx];
-			break;
-	}
-	return result;
-}
+Address param_address(Mode mode, ParamIdx idx);
 
 // clang-format off
 #if LOGGING_ENABLED
@@ -66,12 +83,7 @@ static const char param_names[EUCLID_NUM_PARAMS][PARAM_NAME_LEN] = {
 };
 
 // TODO: This implementation is brittle and will not scale with more modes. 
-bool param_name(char *result, Mode mode, ParamIdx idx) {
-	if (!result) return false;
-
-	memcpy(result, param_names[idx], PARAM_NAME_LEN);	
-	return true;
-}
+bool param_name(char *result, Mode mode, ParamIdx idx);
 
 #endif
 // clang-format on
